task(providerCleanup): Limit StationsProvider logic to handle only the stations by untroducing a new appStore. FWM-34
