+++ package.json
@@ -43,7 +43,8 @@
         "react-leaflet-cluster": "2.1.0",
         "sharp": "0.32.6",
         "tailwindcss": "3.3.3",
-        "typescript": "5.2.2"
+        "typescript": "5.2.2",
+        "zustand": "4.5.0"
     },
     "devDependencies": {
         "@types/leaflet": "1.9.4",
