+++ src/components/Home/HomepageMap.tsx
@@ -1,29 +1,23 @@
 import dynamic from "next/dynamic";
-import { useMemo, useState } from "react";
 import { useStationsProvider } from "@/providers/StationsProvider";
 import L, { MarkerCluster } from "leaflet";
 import { CONFIG } from "@/common/mapSettings";
 import { MarkerCustomAttrs } from "@/types/customMarker";
 import { assetUrl } from "@/common/assetsHandling";
 
+import { useAppStore } from "@/hooks/useAppStore";
 import BaseModal from "@/components/BaseComponents/BaseModal";
 import StationModalContent from "@/components/Home/StationModalContent";
 import MapControls from "@/components/MapControls/MapControls";
 const BaseMap = dynamic(() => import("@/components/BaseComponents/BaseMap"), {
     ssr: false,
 });
-const BaseMarker = dynamic(
-    () => import("@/components/BaseComponents/BaseMarker"),
-    {
-        ssr: false,
-    }
-);
-const MarkerClusterGroup = dynamic(
-    () => import("react-leaflet-cluster"),
-    {
-        ssr: false,
-    }
-);
+const BaseMarker = dynamic(() => import("@/components/BaseComponents/BaseMarker"), {
+    ssr: false,
+});
+const MarkerClusterGroup = dynamic(() => import("react-leaflet-cluster"), {
+    ssr: false,
+});
 
 export default function HomepageMap() {
     const zoomLevel = CONFIG.default_zoom_level;
@@ -32,21 +26,20 @@ export default function HomepageMap() {
     const minZoom = CONFIG.minZoom;
 
     // Create the state for the modal info
-    const [activeStation, setActiveStation] = useState(0);
-    const [markers, setMarkers] = useState<any>([]);
-    const stationsProvider = useStationsProvider();
+    const { setIsStationModalOpen, setActiveStation } = useAppStore();
+    const { stations } = useStationsProvider();
+
     const handleModal = (value: boolean, stationId: number) => {
-        stationsProvider.handleModal(value);
+        setIsStationModalOpen(value);
         setActiveStation(stationId);
     };
-    const isModalOpen = stationsProvider.isStationModalOpen;
 
     const createClusterCustomIcon = function (cluster: MarkerCluster) {
-        const markersInCluster: L.Marker[]  = cluster.getAllChildMarkers();
+        const markersInCluster: L.Marker[] = cluster.getAllChildMarkers();
         const iconsOfCluster: string[] = [];
         markersInCluster.forEach((element) => {
             if (element.options.customAttr) {
-                const { assetId } : MarkerCustomAttrs = element.options.customAttr;
+                const { assetId }: MarkerCustomAttrs = element.options.customAttr;
                 const renderImg = assetUrl(assetId);
                 iconsOfCluster.push(renderImg);
             }
@@ -67,33 +60,21 @@ export default function HomepageMap() {
         });
     };
 
-    useMemo(() => {
-        setMarkers(
-            stationsProvider.stations.map((station) => {
-                return (
-                    <BaseMarker
-                        position={station.location.coordinates.reverse()}
-                        key={station.id}
-                        stationId={station.id}
-                        weatherDescription={
-                            station.accuweather_location
-                                .current_weather_description
-                        }
-                        assetId={station.accuweather_location.weather_condition_icon.asset}
-                        handleClick={handleModal}
-                    />
-                );
-            })
+    const markers = stations.map((station) => {
+        return (
+            <BaseMarker
+                position={station.location.coordinates.reverse()}
+                key={station.id}
+                stationId={station.id}
+                weatherDescription={station.accuweather_location.current_weather_description}
+                assetId={station.accuweather_location.weather_condition_icon.asset}
+                handleClick={handleModal}
+            />
         );
-    }, [stationsProvider.stations]);
+    });
 
     return (
-        <BaseMap
-            zoom={zoomLevel}
-            center={defaultCenter}
-            maxBounds={maxBounds}
-            minZoom={minZoom}
-        >
+        <BaseMap zoom={zoomLevel} center={defaultCenter} maxBounds={maxBounds} minZoom={minZoom}>
             <MarkerClusterGroup
                 chunkedLoading
                 iconCreateFunction={createClusterCustomIcon}
@@ -112,11 +93,8 @@ export default function HomepageMap() {
                 <MapControls />
             </div>
             <div className="absolute bottom-0">
-                <BaseModal isOpen={isModalOpen}>
-                    <StationModalContent
-                        isOpen={isModalOpen}
-                        activeStation={activeStation}
-                    ></StationModalContent>
+                <BaseModal>
+                    <StationModalContent />
                 </BaseModal>
             </div>
         </BaseMap>
